@@ -16,7 +16,7 @@
     <description>Exercise for solving the hierarchy tree issues</description>
 
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <gson.version>2.10</gson.version>
         <commons-lang3.version>3.12.0</commons-lang3.version>
         <commons-collections4.version>4.4</commons-collections4.version>
